apiexam/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==0.19.1
sqlalchemy<=1.4.41
PyMySQL==1.0.2
httpx
pytest
pytest-asyncio
requests
Faker
cryptography